Added Model classes

+++ fullstack-backend/src/main/java/com/HolidayTracker/fullstackbackend/model/User.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 public class User {
 
 
-    private int ID;
+    private long ID;
     private String Email;
     private String Department;
     private String Data;
